Alterar Senha no Perfil

+++ codigo-fonte/Models/MaterialModel.cs
@@ -18,7 +18,7 @@ namespace PUConstruir.Models
         [Required(ErrorMessage = "Digite o Preço Padrão do Material (Se não souber, digite o número zero")]
         public decimal ValorPadrao { get; set; }
 
-        public int Quantidade { get; set; }
+        public decimal Quantidade { get; set; }
 
         public decimal Altura { get; set; }
